pyproject.toml
[project]
name = "pytest-api-cov"
version = "1.3.5"
description = "Pytest Plugin to provide API Coverage statistics for Python Web Frameworks"
readme = "README.md"
authors = [{ name = "Barnaby Gill", email = "barnabasgill@gmail.com" }]
license = { text = "Apache-2.0" }
requires-python = ">=3.10"

dependencies = [
    "pydantic>=2.0.0",
    "rich>=10.0.0",
    "tomli>=1.2.0",
    "pytest>=6.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
django = ["django>=4.0.0"]
fastapi = ["fastapi>=0.68.0", "httpx>=0.20.0", "starlette>=0.14.0"]
flask = ["flask>=2.0.0"]

[project.urls]
Source = "https://github.com/BarnabasG/api-coverage"

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "path>=16.0.0",
    "pytest-cov>=6.2.1",
    "pytest-sugar>=1.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.12.3",
    "typeguard>=4.4.4",
    "vulture>=2.14",
    "types-PyYAML>=6.0",
    "django>=4.0.0",
    "fastapi>=0.68.0",
    "flask>=2.0.0",
    "httpx>=0.20.0",
    "starlette>=0.14.0",
    "pybencher>=2.1.0",
]

# API COVERAGE
[tool.pytest_api_cov]
fail_under = 70
show_covered_endpoints = true
show_uncovered_endpoints = true
show_excluded_endpoints = true
exclusion_patterns = ["xyz"]
report_path = "reports/pytest_api_cov.json"

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration", "example/tests"]
pythonpath = ["src"]

[tool.coverage.report]
exclude_lines = ['if __name__ == "__main__":', "if TYPE_CHECKING:"]
show_missing = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pytest_api_cov"]

[tool.hatch.build.targets.sdist]
include = ["src/pytest_api_cov/", "README.md", "LICENSE", "pyproject.toml"]

# [[tool.uv.index]]
# name = "testpypi"
# url = "https://test.pypi.org/simple/"
# publish-url = "https://test.pypi.org/legacy/"
# explicit = true

[project.entry-points."pytest11"]
pytest_api_cov = "pytest_api_cov.plugin"

[project.scripts]
pytest-api-cov = "pytest_api_cov.cli:main"

[tool.ruff]
src = ["src", "tests", "example"]
line-length = 120
fix = true
lint.select = [
    "E", "W",   # Pydocstyle
    "D",        # Pydocstyle
    "F",        # Pyflakes rules
    "I",        # Sort imports properly
    "C90",
    "N",
    "S",
    "YTT",
    "ANN",
    "ASYNC",
    "BLE",
    "FBT",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "EM",
    "FA",
    "ISC",
    "T20",
    "PYI",
    "PT",
    "Q",
    "RSE",
    "RET",
    "SLF",
    "SLOT",
    "SIM",
    "TID",
    "TCH",
    "INT",
    "ARG",
    "PTH",
    "TD",
    "FIX",
    "ERA",
    "PD",
    "PGH",
    "PL",
    "TRY",
    "FLY",
    "NPY",
    "PERF",
    "RUF",
]
lint.ignore = [
    "D203",
    "D213",
    "COM812",
    "TRY003",
    "EM101",
    "E402",
    "ERA001",
    "FBT001",
    "FBT002",
    "TD002",
    "TD003",
    "FIX002",
    "PLC0415",
    "PLR0912",
    "PLR0915",
    "C901",
    # Print statements are fine for CLI tools
    "T201",
    # Any types are common in plugin/interop code
    "ANN401",
    # Exception patterns
    "EM102",
    # Try-except patterns
    "S110",
    "S112",
    # Magic numbers
    "PLR2004",
    # Private member access
    "SLF001",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",
    "ANN",
    "D100",
    "PLR2004",
    "SLF001",
    "FBT001",
    "PLR0913",
    "C901",
    "N806",
    "D"
]

[tool.mypy]
files = ["src"]
strict = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[tool.vulture]
exclude = []
min_confidence = 80
paths = ["src", "tests", "example"]
sort_by_size = true
verbose = false
